// src/imports/api.ts
export interface ApiPlayerIdGoalsGet200Response {
}
// ...
export interface GetPlayerDTO {
    'id': string;
    'firstName': string;
    'lastName': string;
    'position': string;
    'alternativePositions': Array<string> | null;
    'shirtNumber': ApiPlayerIdGoalsGet200Response;
    'dateOfBirth': string;
    'national': string;
    'isCaptain': boolean;
    'isActive': boolean;
    'team': GetTeamDTO | null;
}
// ...
export interface GetTeamDTO {
    'id': string;
    'name': string;
    'shortName': string;
    'stadium': string;
    'city': string;
    'foundedYear': ApiPlayerIdGoalsGet200Response;
    'coach': string;
    'captain': GetPlayerDTO | null;
    'preferredFormation': string | null;
    'playingStyle': string | null;
}